Add the HD control, tidy the anchors and make motion components list-driven
- HD control beside the zoom buttons, desktop only: every device starts on the
  light preset and the choice is remembered in the browser. No automatic
  frame-rate switching, which is what looked random.
- X-Ray segment uses its own icon instead of the cross-section one.
- Conveyor sensors sit on the belt near the machine, stacked on one line like
  the motor's anchors, instead of at the far end of the conveyor.
- Rotor sensors light both rotors, not just the upper one.
- Leader lines draw critical over warning over normal, re-sorted when an alarm
  starts, so a yellow anchor never covers a red one.
- PartSpinner and PartVibrator replace the hand-placed rotor and shake
  components: a list of parts, each with its own axis, share and phase,
  optionally driven by a sensor reading.
- Clicking a card whose sensor is off screen pulls the camera back just far
  enough to show it.
- History sums in double precision; chart times use the browser's timezone.

+++ Assets/Scripts/Runtime/TelemetryHistory.cs
@@ -78,7 +78,9 @@ public struct TelemetryHistoryBucket
     public int SampleCount;
     public float Min;
     public float Max;
-    public float Sum;
+    // Double: a week-long bucket sums hundreds of thousands of samples, and a
+    // float sum would round small readings away.
+    public double Sum;
     public float Last;
     public float NormalSeconds;
     public float WarningSeconds;
@@ -86,7 +88,7 @@ public struct TelemetryHistoryBucket
     public float UnavailableSeconds;
 
     public bool HasValue => SampleCount > 0;
-    public float Mean => SampleCount > 0 ? Sum / SampleCount : 0f;
+    public float Mean => SampleCount > 0 ? (float)(Sum / SampleCount) : 0f;
     public float CoveredSeconds => NormalSeconds + WarningSeconds + CriticalSeconds + UnavailableSeconds;
     public bool HasData => HasValue || CoveredSeconds > 0f;
